QT:MCU: Robot Following for the Line again

+++ 40_ServiceApplication_QT/LF_ServiceApplication/BleDataManager.h
@@ -129,7 +129,7 @@ signals:
     void BleDatMngrSignal_UpdateErrorWeigthData( float ErrW1,float ErrW2,float ErrW3,float ErrW4,float ErrW5,float ErrW6,float ErrW7,
                                                 float ErrW8,float ErrW9,float ErrW10,float ErrW11,float ErrWM);
 
-    void BleDatMngrSignal_UpdatePidData(float Kp,float Ki,float Kd,float ProbeTime);
+    void BleDatMngrSignal_UpdatePidData(float Kp,float Ki,float Kd,uint32_t ProbeTime);
     void BleDatMngrSignal_UpdateVehCfgData(float ExpectedAvSpd,uint32_t BlinkLedSt, uint32_t TryDetEndLin);
 
     void BleDatMngrSignal_RefreshErrorIndicatorView( uint8_t S0,uint8_t S1,uint8_t S2,uint8_t S3,uint8_t S4,uint8_t S5,
